Refactor KubeConfig into KubeConverter
It turns out we don't need all variables all the time.

* kubeConfig also means 'config for kube'
* with the introduction of job factory 'Convert' only does variables,
  renamed to Variables
* ApplyBPMInfo returns k8s resource specs, does not rely on 'Convert'
  at all, renamed to BPMResources
* explicit inputs and outputs instead of storing results on 'KubeConfig' struct
* split into separate files
* extract jobToContainer funcs into container factory
